+++ source/terms/procedures/ArrivalQualifiers.cs
@@ -5,47 +5,47 @@ namespace Arinc424.Procedures.Terms;
 /// </summary>
 /// <remarks>See table 5-6.</remarks>
 [Flags]
-public enum ArrivalQualifiers : uint
+public enum ArrivalQualifiers : ushort
 {
-    Unknown = 0u,
+    Unknown = 0,
     /// <summary>
     /// DME required.
     /// </summary>
-    DistanceEquipment = 1u,
+    DistanceEquipment = 1,
     /// <summary>
     /// Radar required.
     /// </summary>
-    Radar = 1u << 1,
+    Radar = 1 << 1,
     /// <summary>
     /// RF Leg capability required.
     /// </summary>
-    ConstantRadius = 1u << 2,
+    ConstantRadius = 1 << 2,
     /// <summary>
     /// GNSS required.
     /// </summary>
-    GlobalNavigation = 1u << 3,
+    GlobalNavigation = 1 << 3,
     /// <summary>
     /// Helicopter STAR to Runway.
     /// </summary>
-    Helicopter = 1u << 4,
+    Helicopter = 1 << 4,
     /// <summary>
     /// Continuous Descent STAR.
     /// </summary>
-    Continuous = 1u << 5,
+    Continuous = 1 << 5,
     /// <summary>
     /// VOR/DME RNAV.
     /// </summary>
-    AreaNavigation = 1u << 6,
+    AreaNavigation = 1 << 6,
     /// <summary>
     /// Database supported RNAV.
     /// </summary>
-    DatabaseAreaNavigation = 1u << 7,
+    DatabaseAreaNavigation = 1 << 7,
     /// <summary>
     /// FMS required.
     /// </summary>
-    FlightManagement = 1u << 8,
+    FlightManagement = 1 << 8,
     /// <summary>
     /// Conventional Arrivals.
     /// </summary>
-    Conventional = 1u << 9,
+    Conventional = 1 << 9
 }
